% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamples.R
\name{checkSamples}
\alias{checkSamples}
\title{checkSamples}
\usage{
checkSamples(samples = samples)
}
\arguments{
\item{samples}{Vector of class \emph{character}.}
}
\value{
A \emph{data.frame} with the consistency checks for each element in \emph{samples}.
}
\description{
checks if a shapefile
}
\details{
{Checks if a shapefile - or a set of - contains necessary columns and if these have the right format. It searches for:
\itemize{
 \item{\emph{sampler} - Character vector with name of responsible person.}
 \item{\emph{date} - Date vector with the date on which each sample was collected (formated as "yyyy-mm-dd").}
 \item{\emph{label} - Character vector sample label (e.g. land cover class).}}}
}
\examples{
{}
}
\seealso{
\code{\link{labelCheck}}
}
