% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLabel.R
\name{compareLabel}
\alias{compareLabel}
\title{compareLabel}
\usage{
compareLabel(x, y, x.label, y.label, na.count = 0, max.na.count = 0)
}
\arguments{
\item{x}{Object of class \emph{data.frame} with target profiles.}

\item{y}{Object of class \emph{data.frame} with reference profiles.}

\item{x.lab}{\emph{Character vector} with classes of \emph{x}.}

\item{y.lab}{\emph{Character vector} with classes of \emph{y}.}
}
\value{
A \emph{list}.
}
\description{
Identifies samples that potentially require class label corrections.
}
\details{
{The function cross-correlates \emph{x} and \emph{y}. Then,for each row, the function returns 
the element in \emph{y.label} with the highest correlation. The funal output of the function consists:
\itemize{
 \item{\emph{cross.cor} - Median, minimum and maximum values for each column in \emph{x} over each unique class in \emph{y}.}
 \item{\emph{label.comapre} - \emph{data.frame} showing \emph{y.label} and the nbest match in \emph{y.label}.}
 \item{\emph{na.stats} - \emph{data.frame} showing the count and maximum segment size of NA values for each row in \emph{x}.}
 }}
}
\examples{
{
}
}
\seealso{
\code{\link{extractTS}} \code{\link{phenoCropVal}} \code{\link{phenoCropClass}}
}
