% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{extract2}
\alias{extract2}
\title{extract2}
\usage{
extract2(x, y, x.date, out.date, time.buffer = c(365, 365))
}
\arguments{
\item{x}{A \emph{character} vector with the paths to \emph{RasterLayer} objects or a \emph{list} of \emph{RasterLayers}.}

\item{y}{An object of class \emph{SpatialPoints} or \emph{SpatialPolygons}.}

\item{x.date}{Object of class \emph{Date} with the acquisition dates of each element in \emph{x}.}

\item{out.date}{Object of class \emph{Date} with the desired output dates.}

\item{time.buffer}{Two-element, numeric vector.}
}
\value{
A \emph{list} object.
}
\description{
Extract of values from multi-extent raster objects with a spatial object.
}
\details{
{Creates a rectangular fishnet in a \emph{SpatialPolygons} format based on the 
extent of \emph{x} and the value of \emph{y} which defines the spatial resolution.}
}
