% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAWaR.R
\docType{data}
\name{fieldData2}
\alias{fieldData2}
\title{Point shapefile.}
\format{A SpatialPointsDataFrame}
\usage{
data(fieldData2)
}
\description{
Ground truth data on crop types collected in Uzbekistan derived from the fieldData dataset with poly2sample.
}
\details{
\itemize{
  \item{x}{x coordinates.}
  \item{y}{y coordinates.}
  \item{cover}{Percent overlap between the the pixel each point was derived from and the corresponding polygon.}
  \item{id}{Unique identifier of the polygon the points belong to.}
}
}
\keyword{datasets}
