% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAWaR.R
\docType{data}
\name{fieldDataCluster}
\alias{fieldDataCluster}
\title{Region labels for fieldData}
\format{A data.frame}
\usage{
data(fieldDataCluster)
}
\description{
Output of splitSamples for fieldData.
}
\details{
\itemize{
  \item{region.id}{Region identifier showing which samples are grouped.}
  \item{region.frequency}{Frequency of samples per region.}
}
}
\keyword{datasets}
