% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAWaR.R
\docType{data}
\name{fieldDataTS}
\alias{fieldDataTS}
\title{Samples for fieldData}
\format{
A data.frame
}
\usage{
data(fieldDataTS)
}
\description{
Output of extractTS for fieldData.
}
\details{
\itemize{
  \item{pixel.info}{Unique pixels covered by fieldData with information of x and y coordinates and cover percent.}
  \item{polygon.info}{Percent cover statistics and pixel count per polygon.}
  \item{weighted.mean}{Weighted-mean time-series for each polygon.}
}
}
\keyword{datasets}
