% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBare.R
\name{findBare}
\alias{findBare}
\title{findBare}
\usage{
findBare(x, y)
}
\arguments{
\item{x}{Numeric \emph{vector}.}

\item{y}{Change threshold (in %).}
}
\value{
A \emph{logical} value (TRUE if "bare" and FALSE if not).
}
\description{
Determines if a vector is likely related to a
}
\details{
{Checks if \emph{x} is a likely "bare" time-series. In other 
words, it evaluates if the time series if from a pixel that was likely 
not cultivated assuming that such pixels should have small seasonal 
increase in the NDVI relatively to cultivated pixels. The change threshold 
is defined by \emph{y}. The function starts by normalizing \emph{x} by 
itself avoiding case-specific thresholding. As a consequence, \emph{y} 
is specified as a percent value of change.}
}
\examples{
{

x <- c(293, 770, 1166, 1166, 1562, 2357, 3234, 
5806, 5806, 5678, 5678, 5546, 5536, 5536, 5536, 
5325, 5200, 4726, 3550, 2868, 2365, 2365, 2365)

n <- findBare(x, 100)

}
}
