% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltBit.R
\name{ltBit}
\alias{ltBit}
\title{ltBit}
\usage{
ltBit(path)
}
\arguments{
\item{path}{Acquisition path.}
}
\value{
A \emph{raster} file with the native extention of the input.
}
\description{
Translation of Landsat Collection 1 quality information.
}
\details{
{The function translates Landsat Collection 1 bit quality information into an FMASK equivalent
with labels for Water (1), Cloud Shadow (2),Snow (3), Cloud (4) and No Data (255) and
provides metadata reporting on scene cloud cover. The output file will be names as \emph{"pixel_qa_mask"}}
}
\examples{
\dontrun{

}
}
\keyword{Landsat,}
\keyword{Quality,}
\keyword{Storage}
