% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeLandsat.R
\name{storeLandsat}
\alias{storeLandsat}
\title{storeLandsat}
\usage{
storeLandsat(zpPath, dPath, c1 = TRUE, remove.files = FALSE)
}
\arguments{
\item{zpPath}{Path where the zipped landsat data is stored}

\item{dPath}{Path where the unzipped landsat data should be stored.}

\item{c1}{Is the data from USGS collection 1. Optional logicla argument. Default is TRUE.}

\item{remove.files}{Should the zipped files be deleted after unpacking? Optional logical argument. Default is FALSE.}
}
\value{
A folder structure containing the input Landsat data
}
\description{
Temporal linear interpolation of raster data.
}
\details{
{Local storing of landsat data as downloaded from the ESPA downloading service. A folder
named \emph{LANDSAT}} will be create in the target directory and the output will be split between
two sub-folder:
\itemize{
\item{\emph{"sr"} - Surface reflectances.}
\item{\emph{"infos"} - Metadata.}}
This function unzips the downloaded files and buils a standardized data structure from it with \emph{sr}
(i.e. TILE -> ACQUISITION -> FILES). Additionally, the function translates thebit quality information into
an FMASK equivalent with labels for Water (1), Cloud Shadow (2),Snow (3), Cloud (4) and No Data (255) and
provides metadata reporting on:
\itemize{
\item{\emph{"Date"} - Acquisition date.}
\item{\emph{"Tile"} - Acquisition tile (i.e. path/row.}
\item{\emph{"Path"} - Path to where the data is stored.}
\item{\emph{"Cover"} - Cloud cover percent.}
\item{\emph{"Processed"} - Date when the file was processed.}}
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{storeLandsat}}
}
\keyword{Landsat,}
\keyword{Storage}
