% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCropCycles.R
\name{countCropCycles}
\alias{countCropCycles}
\title{countCropCycles}
\usage{
countCropCycles(x)
}
\arguments{
\item{x}{Target numeric \emph{vector}.}
}
\value{
A \emph{list} with selected indices for \emph{x} and \emph{y}.
}
\description{
Matches two vectors with different lenghts based on their maximum value.
}
\details{
{Uses Dynamic Time Wrapping (DTW) to match \emph{x} and \emph{y}. \emph{z} determines 
the buffer size - expressed in number of data points - used to search for matching records. For 
optimal performance, we recommend that \emph{x} is smoothed a priori.}
}
\examples{
{

x <- c(293, 770, 1166, 1166, 1562, 2357, 3234, 
5806, 5806, 5678, 5678, 5546, 5536, 5536, 5536, 
5325, 5200, 4726, 3550, 2868, 2365, 2365, 2365)

n <- countCropCycles(x)

}
}
