% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBare.R
\name{findBare}
\alias{findBare}
\title{findBare}
\usage{
findBare(x, y)
}
\arguments{
\item{x}{A numeric \emph{vector}.}

\item{y}{A numeric element.}
}
\value{
A logical element.
}
\description{
Shows is a vector is related to bare/fallow land..
}
\details{
{The function computes the amplitude of \emph{x} and returns TRUE if this value is below the threshold given by \emph{y}.}
}
\examples{
{

# vectors to test
x1 <- c(0.1,0.2,0.1,0.2,0.4,0.6,0.8,0.4,0.1) # simulated, crop profile
x2 <- c(0.1,0.3,0.2,0.2,0.1,0.3,0.1,0.2,0.2) # simulated, non-crop profile

# compare profiles
plot(x1, type="l")
lines(x2, col="red")

findBare(x1, 100) # returns TRUE
findBare(x2, 100) # returns FALSE

}
}
