% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBare.R
\name{findBare}
\alias{findBare}
\title{findBare}
\usage{
findBare(x, y)
}
\arguments{
\item{x}{Numeric \emph{vector}.}

\item{y}{Change threshold (in %)}
}
\value{
A logical element.
}
\description{
Determines if a vector is likely related to unused non-croped areas.
}
\details{
{Judges if \emph{x} is a time series of non-agricultural land (poitnetial bare land) based on its amplitude.}
}
