% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meStack.R
\name{meStack}
\alias{meStack}
\title{meStack}
\usage{
meStack(x, y, z, mask = NULL, plot = FALSE, fun = mean)
}
\arguments{
\item{x}{A character vector containing the paths to the images to stack.}

\item{y}{A spatial object from which an extent can be derived.}

\item{z}{Object of class \emph{Date} with the acquisition date for each element in \emph{x}.}

\item{mask}{Object of class \emph{RasterLayer} with the same extent as \emph{y}.}

\item{plot}{Logical argument.}

\item{fun}{A function. Default is mean.}
}
\value{
A list containing a \emph{RasterStack} and related statistics.
}
\description{
Stacking of raster layers with different extents
}
\details{
{The function stacks the raster objects specified in \emph{x}. For each element in \emph{x}, the function crops it by the
extent of \emph{y} and, if their extents differ, fits the extent of \emph{x} to the one of \emph{y}. All new pixels are set to NA. If
\emph{z} is provided, the function will then aggregate all bands acquired in the same date \emph{plot} is set to TRUE, the function will derive basic statistics for each band (i.e. min, max, mean, sd) as well as a plot showing
then mean, min and max for each band. If \emph{mask} is provided, the plot will be based on all non-NA pixels. The final output of the
function is a list containing:
\itemize{
 \item{\emph{stack} - \emph{RasterStack} object.}
 \item{\emph{statistics} - Statistics for each band in the output \emph{RasterStack}.}
 \item{\emph{plot} - Plot showing the mean, minimum and maximum values per band.}
 \item{\emph{control} - Logical vector showing which elements in \emph{x} where used to build the \emph{RasterStack}.}}}
}
\examples{
{}
}
